TrendFactor.py hand coded 50% match with stata, the rest is a bug in the stata code
The bug is line 71 of TrendFactor.do, which says
`bys time_avail_m: keep if _n == 1  // Current dataset is firm-time-level but we only need time-level here`
this is meant to convert the redundant permno-time_avail_m regression results into time_avail_m regression results. the problem is that some of the permnos have missing regression results (typically when they have missing returns). so this randomly drops some months' regression results. the fix is to use keep if _Nobs != . before the _n == 1 filter.